Extract JIT tests from test_functional to the dedicated test module
